ASTImporter fix for recursive import of function
In some circumstances it is possible that a FunctionDecl is
imported duplicatedly during ASTNodeImporter::VisitFunctionDecl.
Checks are added to prevent overwrite of existing imported
FunctionDecl. New assertions are added to check for this
problem too. VisitStaticAssertDecl was updated with same check
because it was discovered that the problem happens here too.

Added tests that provokes this case (ASTImporter and unittests).