ASTImporter::Imported - VisitClassTemplateDecl assertion fix
Fix for assertion
"Try to import an already imported Decl"
that occurs through calling
ASTNodeImporter::VisitClassTemplateDecl.

The problem is related to import of a CXXRecordDecl that is part
of a template ("template describing record"). If the describing
record is imported (on its own, not as part of the described
template) and this is already existing it is not found and
imported redundantly. When the described template is imported
(already during import of the describing record) the existing
template is found and the assertion is triggered.