[ASTImporter] Import friend class templates better.
Do not return an existing Decl when importing friend template.
(If a Decl is a child of a FriendDecl that same Decl can not
exist in another place in the AST.)
Do not import first the definition instead of decl at prototype for
friend class template.
(The ClassTemplateDecl under a FriendDecl can not be replaced with
its definition.)