[ASTImporter] Import implicit functions.
If a CXXRecordDecl with definition is imported and there is an existing
definition, update the definition with implicit functions from the
imported one.
The implicit functions may not be present in the existing definition,
if the existing code did not make use of them. But the imported code may
have references to these functions.