[ASTImporter] Getting index of field should load Decls.
At equality check of fields without name the index of fields is compared.
At determining the index of a field all fields of the parent context
should be loaded from external source to find the field at all.
It is possible that when a field is imported it is not loaded yet into its
DeclContext.