[ASTImporter] Changed unsupported node handling
The previous 'unsupported node" flag is changed to "unsupported construct".
This indicates if there was an unsupported construct (node or other not
importable case) encountered during import.
Normally every kind of AST node should be supported, unless explicitly
marked as not supported. Therefore the generic import functions contain
an assert to find (and implement) not yet supported nodes.