[ASTImporter] Fix import of static and anonymous functions
We must not connect a static function declaration with an extern
function declaration even if they have the same name.  This is true for
functions from anonymous namespaces if they are coming from different
TUs. However we must connect functions from anon namespaces with same
names and type if they are coming from the same TU.