Add/remove from the Lookuptable consistently (#592)
* Add/remove from the Lookuptable consistently

Add/remove from the Lookuptable in consistency with ImportedFromDecls.
When we map a decl in the 'to' context to something in the 'from'
context then and only then we add it to the lookup table.  When we
remove a mapping then and only then we remove it from the lookup table.

* Add comment to MapImported